Move all TestingFilterFunctions to OrcTester
Move TestingWithInputIntegerFilterFunction and
TestingNoInputFilterFunction to OrcTester.from TestSelectiveOrcReader.
Move TestingFlatMapFilterFunction to OrcTester from
TestMapFlatSelectiveStreamReader as well.

Just a little cleanup, nothing really too important here.
